+++ docs/changelog.md
@@ -14,7 +14,8 @@ Includes various usability enhancements identified since the region editor was l
 #### Changelog
 
  - [improvement] Add `learning-rate` config option to adjust how the background model is updated
- - [bugfix] The region editor will now ask before quitting if there are unsaved changes
+ - [bugfix] The region editor now always prompts on any unsaved changes
+ - [general] The region editor no longer prompts for a save path if `-s`/`--save-region` was specified
 
 ### 1.6 (October 15, 2023)